// src/services/models/company-role-structure.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { CompanyWorker } from './company-worker';

/**
 * 
 * @export
 * @interface CompanyRoleStructure
 */
export interface CompanyRoleStructure {
    /**
     * 안전보건조정자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'coordinateOfficer': Array<CompanyWorker>;
    /**
     * 안전보건총괄책임자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'headOfficer': Array<CompanyWorker>;
    /**
     * 보건관리자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'healthManager': Array<CompanyWorker>;
    /**
     * 산업보건의
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'industrialHealthOfficer': Array<CompanyWorker>;
    /**
     * 사업주
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'owners': Array<CompanyWorker>;
    /**
     * 안전보건관리담당자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'safetyHealthManager': Array<CompanyWorker>;
    /**
     * 안전관리자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'safetyManager': Array<CompanyWorker>;
    /**
     * 안전보건관리자책임자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'safetyOfficer': Array<CompanyWorker>;
    /**
     * 관리감독자
     * @type {Array<CompanyWorker>}
     * @memberof CompanyRoleStructure
     */
    'supervisor': Array<CompanyWorker>;
}