// src/services/models/company-worker.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CompanyWorker
 */
export interface CompanyWorker {
    /**
     * 
     * @type {string}
     * @memberof CompanyWorker
     */
    'department'?: string;
    /**
     * 
     * @type {string}
     * @memberof CompanyWorker
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof CompanyWorker
     */
    'role'?: string;
}